Introduce isExpanded method to clarify drawer state logic.
The new isExpanded method is equals to the logical negation of the
legacy isCollapsed method; all usages of the legacy isCollapsed method
are updated accordingly.

The isCollapsed method is updated to return true only when the drawer is
fully collapsed.

With these two method the drawer can either be:
* expanded -- only isExpanded() returns true;
* neither expanded nor collapsed -- both methods return false;
* collapsed -- only isCollapsed retturns true.

Bug: 404593897
Test: manua functinality smoke test
Flag: EXEMPT refactoring
Change-Id: I3f37ebd37276b4ef5550b122f93a3812e8fa4979