Add options to allow Activity Starts (BAL)
Opt in 2 PendingIntents and one IntentSender to allow background activity starts (BAL).
The IntentSender is the likely culprit causing Photos to be unable to copy a link (b/424827362).
The assumption here is that all 3 actions are triggered by the user choosing an action (which makes the IntentResolver responsible to opt in).

Flag: EXEMPT bugfix
Bug: 424827362
Test: atest IntentResolver-tests-unit
Change-Id: Ia47a47843ff77a97fc223dd2d7eef711f5491baa