Implement basic desktop UI for Chooser
This change introduces a freeform window Chooser UI variant -- a
centered dialog.

The mode is enabled when Chooser is launched under the following
conditions:
  (a) In the regular (i.e., non-interactive) mode.
  (b) The activity is in multi-window mode.
  (c) The device has the freeform window feature.

Notably, this means that the Chooser will still be presented as a bottom
sheet when launched on a desktop device but in full-window mode.

Known issues:
 - For the desktop mode, the modify share option is still displayed in
   the header, not as a custom action.
 - Various UI details need refinement.

Code changes:
 - Created DrawerUiModeController to manage switching between
   bottomsheet and dialog modes, as well as handling system window
   insets changes.
 - Updated ResolverDrawerLayout to support height "wrap_content" spec.

Bug: 453752027
Test: Manual testing on a desktop device.
Flag: com.android.intentresolver.desktop_ui
Change-Id: Ic21f27168136ad64228abf06013f0b4442d51373