Refactor: Move top-K sorting to ChooserListController
Move the responsibility of selecting the top K ranked targets from
ChooserListAdapter to ChooserListController.

The number of targets to rank is now determined by the mMaxTargetsPerRow
value from ChooserActivity, ensuring the controller has the necessary
information to perform the partial sort. This centralizes the sorting
logic within the controller.

The sortComponents method in ResolverListAdapter is now private as it's
an internal implementation detail.

Bug: 459573554
Test: functionality smoke test
Test: atest IntentResolver-tests-unit
Test: atest IntentResolver-tests-activity
Flag: EXEMPT PURE_REFACTOR
Change-Id: I84cd8a7d08d3daa0065203b9411d3c685827b927